feat: implement GitHub, Trello, and Jira project fetching and task import API routes
